requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
tzdata==2025.1
psycopg[binary]==3.2.9
psycopg-pool==3.2.6
openai>=1.40.0
